requirements.txt
# Django and REST Framework
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-spectacular==0.27.2

# CORS handling
django-cors-headers==4.3.1

# Bitcoin RPC
python-bitcoinrpc==1.0

# Environment variables
python-decouple==3.8

# Database
psycopg2-binary==2.9.9  # For PostgreSQL (optional)

# Caching
redis==5.0.1
django-redis==5.4.0

# Testing
pytest==7.4.4
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
flake8==7.0.0
black==24.2.0
isort==5.13.2

# Security
django-ratelimit==4.1.0
django-extensions==3.2.3

# Monitoring and logging
sentry-sdk==1.40.6

# Development tools
django-debug-toolbar==4.2.0
ipython==8.21.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
celery==5.3.6
kombu==5.3.5